backend/requirements.txt
Flask
Flask-MySQLdb
mysql-connector-python
mysqlclient
Flask-Cors
requests
minio
werkzeug